+++ src/copia/cli/commands/__init__.py
@@ -1,2 +1,4 @@
-from .list import list_profile_app as list_app
+from .list import list_command
 from .init import init_command
+from .tui import tui_command
+from .run import run_command