+++ src/copia/cli/config/globals.py
@@ -1,5 +1,5 @@
 """All the CONSTANTS shared by the copia config module"""
-from typing import Literal
+from typing import Literal, TypeAlias
 from pathlib import Path
 
 from typer import get_app_dir
@@ -11,6 +11,6 @@ LOCAL_COPIA_FILE = Path(".copia.toml")
 GLOBAL_COPIA_FILE = Path(APP_DIR) / "profiles.toml"
 PORT_MIN_VAL = 1
 PORT_MAX_VAL = 65535
-SUPPORTED_ADAPTERS = Literal['mysql', 'postgres']
-CONFIG_SCOPES = Literal["local", "global"]
+Adapter: TypeAlias = Literal['mysql', 'postgres']
+ConfigScope: TypeAlias = Literal["local", "global"]